requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
networkx==3.4.2
sqlparse==0.5.3
sqlglot==26.3.2
google-generativeai==0.8.3
chromadb==0.5.23
qdrant-client==1.12.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
alembic==1.14.0
